third_party/WebKit/LayoutTests/typedcssom/cssKeywordValue.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>

test(function() {
  assert_equals(new CSSKeywordValue('initial').toString(), 'initial');
  assert_equals(new CSSKeywordValue('center').toString(), 'center');
  assert_equals(new CSSKeywordValue('customLemon').toString(), 'customLemon');
  assert_equals(new CSSKeywordValue(' Hello World').toString(), CSS.escape(' Hello World'));
  assert_equals(new CSSKeywordValue('3').toString(), CSS.escape('3'));
}, 'toString() returns a string with a format similar to CSS.escape. This test also ' +
    'implies that toCSSValue supports all keywords including custom identifiers');

test(function() {
  let keywordValue = new CSSKeywordValue('foo');
  assert_equals(keywordValue.value, 'foo');

  keywordValue.value = 'bar';
  assert_equals(keywordValue.value, 'bar');
}, 'The keyword used in KeywordValue can be updated');

</script>
<body>
</body>
